pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "saas-core"
version = "0.1.0"
description = "Core authentication library for Hub & Spoke SaaS architecture with Google Cloud Identity Platform integration"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["authentication", "jwt", "gcip", "google-cloud", "firebase", "saas"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Security :: Authentication",
]

dependencies = [
    "pyjwt[crypto]>=2.8.0",
    "cryptography>=41.0.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
database = [
    "sqlalchemy[asyncio]>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/saas-core"
Documentation = "https://github.com/yourusername/saas-core#readme"
Repository = "https://github.com/yourusername/saas-core"
Issues = "https://github.com/yourusername/saas-core/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["saas_core*"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
